refactor relevant tools into conversion for now, create a new _data_corrections file, start esgf data attrs updater json
